xen/livepatch/x86/arm32: Force .livepatch.[xen].depends section to be uint32_t aligned.
By default when using objcopy we lose the alignment when we copy it from xen-syms -
with the result that alignment (on ARM32 for example) can be 1:

  [Nr] Name              Type            Addr     Off    Size   ES Flg Lk Inf Al
..
  [ 6] .livepatch.depend PROGBITS        00000000 000093 000024 00   A  0   0  1

That, combined with wacky offset means it will be loaded in
memory with the wrong alignment:

(XEN) livepatch.c:425: livepatch: xen_bye_world: Loaded .livepatch.depends at 000a08043

and end up crashing on ARM32

This fix forces all the test-cases to be built with a
.livepatch.depends (and livepatch.xen.depends) structure containing the build-id extracted
from the hypervisor (except the xen_bye_world test-case).

We use the 'mkhex' tool instead of 'xxd' as the end result is an 'unsigned'
instead of 'char' type array - which naturally forces the alignment to be of four.
Also the 'mkhex' tools allows us to pass the section name as parameter.

The end result is much better alignment:

  [ 7] .livepatch.depend PROGBITS        00000000 000094 000024 00   A  0   0  4

Note: We also had to fix up the xen_no_buildid testcase as it copied
a xen_nop test-case (which had the .livepatch.xen.depends), so to make it
xen_no_buildid testcase fail, we ended up removing the .livepatch.xen_depends section.

Signed-off-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>