livepatch/arm/x86: Strip note_depends (and note_xen_depends) symbol from test-cases.
Otherwise each livepatch has a global:

    30: 00000000     1 OBJECT  GLOBAL HIDDEN     7 note_depends

which will cause collision when loading.

The solution to this is to remove the symbol via the --strip-symbols.

However that fails as note_depends is in use by .rel.debug_info:
Relocation section '.rel.debug_info' at offset 0x151c contains 113 entries:
 Offset     Info    Type            Sym.Value  Sym. Name
..
00000625  00001e02 R_ARM_ABS32       00000000   note_depends

And the solution to that is to also slap on --strip-debug which removes
various .debug* sections (which livepatch ignores anyhow):
.debug_aranges, .debug_info, .debug_abbrev, .debug_line, .debug_frame,
.debug_str, and their .rel.* sections. And that will remove that.

Alternatively we could use --localize-symbol so that note_depends
is not globally visible.

Signed-off-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>